subsys: net: lib: aws_jobs: Change enum and bugfix for topic length
Changes enum to have prefix so that we don't reserve the name space for
`IN_PROGRESS` also updated the topic lengths not to includ the `\0`
delimiter of the string. Since this would cause an error in some cases
and the mqtt library handles the strings as utf-8 with fixed sizes.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>