samples: bootloader: Change bootloader to account for mcuboot padding
This change alters the bootloader to check for mcuboot image header data
before trying to boot any of the images. After header data is found it
will compare the image version numbers and boot the highest numbered
version. It will also check the image being booted with the metadata and
signature from b0.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>