remove excess ifdef.
move de-initialization of ctx variable

move check of length of buffer

fix incorrect define checked in dfu_target_img_type

The variables needs to exist since its the linker that removed
the unused code not the compiler. So we get a compiler error
if we have not set either of the DFU targets at the moment.

add check when no targets selected

check if non-zero instead of if negative

add nullcheck

add unit test

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>