mgmt: fmfu: Add callback to fmfu init function
Add callback so that the user decides what needs to be done before
starting the full modem update process when receiving a full modem
update command on SMP.

The user will have to shutdown the modem and initialize it in DFU_MODE
as described in the documentation of the library. Since we have no
way of checking which mode the library currently is in we can not just
return an error value. As found in ref. NCSDK-9120 having the modem
library not initialized in DFU_MODE will cause the device to reset.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>