mcuboot: external_flash: Support SPI devices with partition manager
Adds support for the combination of using a `chosen` DTS node and the
KConfig `PM_EXTERNAL_FLASH_MCUBOOT_SECONDARY` to enable using external
flash as the secondary slot in MCUBoot with SPI devices that supports
`jedec` and the `spi-nor` driver.

In our devices this is currently only `nrf9160dk_nrf9160@0.14.0` and up.
Where the `mx25r64` is configured to be connected to `spi3` in the
device tree.

Ref. NCSDK-11451

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>