pm: Always put the rpmsg_nrf53_sram partition inside sram_nonsecure
Always put the rpmsg_nrf53_sram partition inside sram_nonsecure
instead of doing it only when CONFIG_BT_RPMSG_NRF53 is enabled.

This option is not always enabled so we can't do it conditionally on
it.

missing partitions are silently ignored so it is safe to span it
unconditionally.

This fixes a boot failure in the openthread/cli sample when TF-M is
enabled.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>