subsys: partition_manager: Fix pcd_sram placement for samples
Both #6137 and #6861 failed to work for both use cases.

This patch ensures that `pcd_sram` is placed in the secure partition
when Trustzone is used. If it not `pcd_sram` is placed at the end of
RAM.  If Trustzone is used `pcd_sram` is placed at the end of the secure
RAM partition.

In the case of nRF5340, Bluetooth subystem creates
rpmsg_nrf53_sram partition in SRAM for the HCI transport
between the two cores. If the partition is present,
sram_primary and sram_nonsecure partitions are reduced and
the rpmsg_nrf53_sram is not accessible from the non-secure
firmware - an application built with CONFIG_BT=y causes
the secure fault.

Make sram_nonsecure partition, which defines the non-secure
SRAM, span rpmsg_nrf53_sram as well

Ref. KRKNWK-12196
Ref. NCSDK-13647

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>